Feat: Task 2 - Authentication Module and Documentation
It was added the SignOut functionality which leads the user to the log in screen after signing out. Temporarily, it was added an action within the menu to trigger this fuction.

Documentation:
It was fixed the "How to implement it" section of the UserAuthenticationModule.md file in order explain in the point 3 that the mUserAuthentication has to be declared as a member variable.
Besides it was documented how to sign out.